+++ digiwf-spring-cloudstream-utils-starter/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/utils/configuration/StreamingConfiguration.java
@@ -1,6 +1,6 @@
 package io.muenchendigital.digiwf.spring.cloudstream.utils.configuration;
 
-import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.FunctionRouter;
+import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.RoutingCallback;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.function.context.MessageRoutingCallback;
@@ -18,7 +18,6 @@ public class StreamingConfiguration {
 
     @Bean
     public MessageRoutingCallback customRouter() {
-        final FunctionRouter functionRouter = new FunctionRouter(streamingProperties.getTypeMappings());
-        return functionRouter.customRouter();
+        return new RoutingCallback(this.streamingProperties.getTypeMappings());
     }
 }
